pipeline/requirements.txt
yfinance==1.2.0
pandas==2.2.3
numpy==2.2.4
statsmodels==0.14.4
supabase==2.15.1
python-dotenv==1.1.0
joblib==1.4.2
requests==2.32.3
lxml==5.3.0
beautifulsoup4==4.13.3
tqdm==4.67.1
